+++ internal/app/wwctl/container/imprt/main.go
@@ -1,7 +1,9 @@
 package imprt
 
 import (
-	"github.com/hpcng/warewulf/internal/pkg/api/container"
+	"github.com/hpcng/warewulf/internal/pkg/container"
+
+	apicontainer "github.com/hpcng/warewulf/internal/pkg/api/container"
 	"github.com/hpcng/warewulf/internal/pkg/api/routes/wwapiv1"
 	"github.com/spf13/cobra"
 )
@@ -13,6 +15,9 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 	if len(args) == 2 {
 		name = args[1]
 	}
+	if list, _ := container.ListSources(); len(list) == 0 {
+		SetDefault = true
+	}
 
 	cip := &wwapiv1.ContainerImportParameter{
 		Source:   args[0],
@@ -23,6 +28,7 @@ func CobraRunE(cmd *cobra.Command, args []string) (err error) {
 		Default:  SetDefault,
 		SyncUser: SyncUser,
 	}
-	_, err = container.ContainerImport(cip)
+
+	_, err = apicontainer.ContainerImport(cip)
 	return
 }