+++ CHANGELOG.md
@@ -127,6 +127,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update iPXE building script
 - Send Info, Recv, Send, and Out messages to stdout; and others to stderr
 
+- default is now protected and can't be deleted
+- first container imported container is added to the default profile
 ## [4.4.0] 2023-01-18
 
 ### Added